tests Read submitted JSON file.

+++ tests/submit.php
@@ -11,31 +11,68 @@
       $theme->main_begin();
         $theme->article_begin("Tests Submission");
 
-          // Clear our test tables
-          $result = $util->db->Remove("test_project");
-          $result = $util->db->Remove("test_target");
-          $result = $util->db->Remove("test_result");
+          if ( $_SERVER['REQUEST_METHOD'] === 'POST' ){
+           if (
+              /*(($_FILES["file"]["type"] == "image/gif")
+              || ($_FILES["file"]["type"] == "image/jpeg")
+              || ($_FILES["file"]["type"] == "image/pjpeg"))
+              &&*/ ($_FILES["file"]["size"] < 20000)
+            ) {
+              if ($_FILES["file"]["error"] > 0) {
+                echo "Return Code: " . $_FILES["file"]["error"] . "<br />";
+              } else {
+                echo "Name: " . $_FILES["file"]["name"] . "<br />";
+                echo "Type: " . $_FILES["file"]["type"] . "<br />";
+                echo "Size: " . ($_FILES["file"]["size"] / 1024) . " Kb<br />";
+                echo "Temp file: " . $_FILES["file"]["tmp_name"] . "<br />";
+                $fileText = $_FILES["file"]["tmp_name"];
 
-          // Add our new entries to the test tables
-          $result = $util->db->Add("test_project", "project_name", "'project'");
-          $project_id = $util->db->GetLastInsertID();
-
-          $theme->article_addline("project_id=$project_id");
 
 
-          $result = $util->db->Add("test_target", "target_name, target_projectid", "'target', '$project_id'");
-          $target_id = $util->db->GetLastInsertID();
+                echo "File=" . $fileText . "\n";
+                echo "Contents=" . file_get_contents($fileText) . "\n";
+                echo "Decoding JSON file" . "\n";
+                var_dump($util->JSONDecode(file_get_contents($fileText)));
 
-          $theme->article_addline("target_id=$target_id");
 
 
-          $result = $util->db->Add("test_result", "result_name, result_state, result_targetid", "'result', 'passed', '$target_id'");
-          $result_id = $util->db->GetLastInsertID();
+                // Clear our test tables
+                $result = $util->db->Remove("test_project");
+                $result = $util->db->Remove("test_target");
+                $result = $util->db->Remove("test_result");
 
-          $theme->article_addline("result_id=$result_id");
+                // Add our new entries to the test tables
+                $result = $util->db->Add("test_project", "project_name", "'project'");
+                $project_id = $util->db->GetLastInsertID();
 
-          if ($result) $theme->article_addline("Succeeded");
-          else $theme->article_addline("Failed");
+                $theme->article_addline("project_id=$project_id");
+
+
+                $result = $util->db->Add("test_target", "target_name, target_projectid", "'target', '$project_id'");
+                $target_id = $util->db->GetLastInsertID();
+
+                $theme->article_addline("target_id=$target_id");
+
+
+                $result = $util->db->Add("test_result", "result_name, result_state, result_targetid", "'result', 'passed', '$target_id'");
+                $result_id = $util->db->GetLastInsertID();
+
+                $theme->article_addline("result_id=$result_id");
+
+                if ($result) $theme->article_addline("Succeeded");
+                else $theme->article_addline("Failed");
+              }
+            } else {
+              $theme->article_addline("Invalid file");
+            }
+          } else {
+            $theme->article_addline("<form enctype=\"multipart/form-data\" action=\"submit.php\" method=\"post\">");
+            $theme->article_addline("  <input type=\"hidden\" name=\"MAX_FILE_SIZE\" value=\"30000\" />");
+            $theme->article_addline("  Upload this file: <input type=\"file\" name=\"file\" />");
+            $theme->article_addline("  <input type=\"submit\" value=\"Submit File\" />");
+            $theme->article_addline("</form>");
+          }
+
         $theme->article_end();
       $theme->main_end();
